Certificate manager and ref counted certificate type (#3688)
* New certificate and secrete abstraction with certificate infrastructure and certificate/secret stores. 
* leak testing for certificate resources to validate new ref counting in every test assembly
* Migration guide updated and documentation added for all
* Remove IDisposable from Nodestate and UserIdentity.
* UserCertificateTokenHandler now just points into certificate registry/stores, no dispose required
* Fix CA2000 warnings and cleanup of other diagnostics, increase warning levels in .editorconfig

+++ Applications/McpServer/Tools/NodeManagementServiceTools.cs
@@ -90,9 +90,11 @@ namespace Opc.Ua.Mcp.Tools
                         DataType = DataTypeIds.BaseDataType,
                         AccessLevel = AccessLevels.CurrentReadOrWrite,
                         UserAccessLevel = AccessLevels.CurrentReadOrWrite,
-                        SpecifiedAttributes = (uint)(NodeAttributesMask.DisplayName |
-                            NodeAttributesMask.DataType | NodeAttributesMask.AccessLevel |
-                            NodeAttributesMask.UserAccessLevel)
+                        SpecifiedAttributes =
+                            (int)NodeAttributesMask.DisplayName |
+                            (int)NodeAttributesMask.DataType |
+                            (int)NodeAttributesMask.AccessLevel |
+                            (int)NodeAttributesMask.UserAccessLevel
                     }),
                     _ => new ExtensionObject(new ObjectAttributes
                     {
